+++ examples/fe-fpm-cli/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @sap-ux/fe-fpm-cli
 
+## 1.0.32
+
+### Patch Changes
+
+#### Workspace Updates
+
+- @sap-ux/fe-fpm-writer 1.3.12 → 1.3.13
+
+## 1.0.31
+
+### Patch Changes
+
+#### Workspace Updates
+
+- @sap-ux/fe-fpm-writer 1.3.11 → 1.3.12
+
+## 1.0.30
+
+### Patch Changes
+
+#### Workspace Updates
+
+- @sap-ux/fe-fpm-writer 1.3.10 → 1.3.11
+
 ## 1.0.29
 
 ### Patch Changes